Fix to not reinstall frameworks when .git/HEAD is a ref
Summary: Before running a framework, we check to make sure the git
hash hasn't changed. If the .git/HEAD file is a ref instead of a hash,
this check always fails, resulting in us needlessly deleting and reinstalling
a framework on each test run. Fix this by detecting if .git/HEAd is a ref
and loading the hash out of that ref, stopping us from wasting some time
and CPU on reinstalling.

Reviewed By: @fredemmott

Differential Revision: D2329855