Do naming of instanceof in naming phase
Summary: No point doing the naming over and over again in typing. @int3 wrote
some pretty silly code a year ago...

This also allows obj gets / class gets / function calls to be used on
the RHS of an instanceof expression, and raises an error if we are using
`instanceof $obj->mem` where `mem` is a string or other unknown type.

Reviewed By: @dlreeves

Differential Revision: D2340925