Fix crash/leak in ext_curl
Summary: ext_curl was relying on std::move of a folly::Optional to free
the source of the move. The move constructor is guaranteed to do
so, but std::move can be elided. eg

  void foo(X&& x);
  void bar(X& x) {
    foo(std::move(x)); // not required to call X::X(&&)
  }

By introducing a local variable, this forces the move
constructor to be called (and adds an assert for good
measure).

Reviewed By: @edwinsmith

Differential Revision: D2347237