Add a .declvars directive to assembler
Summary: The hhas assembler currently assigns IDs to variables when the variables are first seen in an instruction. Sometimes (like when implementing closures), more control is needed over the assignment, so add a .declvars directive that takes a list of variables and assigns them IDs. Have the disassembler print out a .declvars for all the non-param named values, for roundtripping purposes.

Reviewed By: @markw65

Differential Revision: D2342134