Port strtr from zend 5.6.10
Summary: Addresses https://github.com/facebook/hhvm/issues/5471

strtr() invoked with an array $replace_pairs as the second argument appears to
perform much worse under HHVM than PHP5.

PHP's implementation of multiple-string search was formerly quite naïve, and it
was the subject of PHP bug 63893. In 2013, it was rewritten to utilize a clever
and efficient algorithm for multi-pattern searching (Wu-Manber), and then
further optimized. It looks like HHVM's implementation is based on the naive
implementation PHP sported before PHP 5.4, which explains the poor performance.

This patch ports over the strtr implementation in php 5.6.10.
Closes https://github.com/facebook/hhvm/pull/5517

Reviewed By: @jwatzman

Differential Revision: D2178123