Remove more of scan-methods
Summary: D2342407 removed the code references, but left the build system
wiring in place, which is kind of annoying to maintain. Nuke all of that
too. I left the actual shell script to generate it in place since I
don't know if you still want that, or if it does something else?

Reviewed By: @​bnell

Differential Revision: D2346840