filter_var_array: do not fall back to FILTER_DEFAULT for all failures
Fallback to `FILTER_DEFAULT` for failed validations of individual
elements was introduced in 2f48df4 while fixing a deeper bug where HHVM
always fell back to FILTER_DEFAULT when any `filter_var()` validation
failed. This fallback behavior is not compatible with the PHP5 engine
which applies no special result processing for failed validations. PHP5
instead only falls back to use of `FILTER_DEFAULT` only when the filter
passed for a given variable is not a valid filter (eg -1).

Fixes #5836