Don't generate an ENABLE_EXTENSION_* variable for required extensions
You can't disable required extensions to begin with, so generating these variables, which is only done to make it easy for the user to set them, is useless, as the user will just get an error if they try to disable it.

Due to where the change is, #5872 (D43749) will conflict with this if that PR is merged first. If this PR is merged before that one, there shouldn't be any conflicts when merging, although it will depend on how picky git decides to be.