typing TypeStructure<T>
Summary: typing TypeStructure<T> as an alias of a shape, with the exception of the classname field, which we make non-nullable when the type parameter T is of a class kind.

(The scope of this diff is lessened due to the lack of typedef expansion support in the runtime. See D2310657 for a mocking of more full-fledged typing of the shape representation of type T)

Reviewed By: @dlreeves

Differential Revision: D2339718