Check Type Parameter Constraints on Declaration Types
Summary: Hack was failing to check if the types used to instantiate a class type
parameter satisified the constraints on that type in various declaration
locations. This means the following code would type check even though it
violates the constraints specified by the class.

```
class Foo<T as num> {
  // Wait Foo of string isn't valid?
    public ?Foo<string> $foo = null;
}
```

This seems to have been broken since at least December 2014. I didn't bother looking at further to track down exactly when this regression was introduced.

To fix this I hooked up constraint checking of `Happly` hints in `NastCheck` as part of the `hint` checks we do. The `NastCheck`s are ran right before we type a class/function so this is the earliest we can perform these checks.

One thing to note is when we are doing `NastCheck.typedef` we need to substitue `Tany` for type parameters declared by the typedef in any `Happly` we are checking because typedefs are not allowed to have constraints.

Reviewed By: @dabek

Differential Revision: D2345338

Conflicts:
	hphp/hack/src/h2tp/test/convert_hack_test_inputs_base.py