Exclude util/light-process.cpp under MSVC
LightProcess is heavily dependent upon fork, which is not available under Windows, so I've taken the route of simply not porting it to MSVC, and re-writing any required code that uses it, to not use it.
This simply removes the sources for them from the build under MSVC, and it shouldn't effect any other platforms.