Handle a few places where MSVC chokes on the atomic loads
Basically the issue here is that MSVC doesn't like it when you use an `std::atomic` with a type other than `void*` as a condition to an `if` statement, so this works around it by just loading the value explicitly.

This is dependent upon an issue with how [folly#264](https://github.com/facebook/folly/pull/264) was merged being resolved first.