Add a testpilot mode to hphp/test/run
Summary: We need a way to tell hphp/test/run that it's okay to emit
some sort of skip since testpilot needs positive acknowledgement
to differentiate a bad test from a test that skipped. Add a new mode
that emits a 'not_relevant' skip type which can be parsed and swallowed
by testpilot.

Reviewed By: @paulbiss

Differential Revision: D2348384