Add macro for HHVM OSS push safety
Summary: This is a temporary workaround for the break
introduced by D2327946 to HHVM. The problem there
was that all the deps are not updated.
This adds HHVM_SECURE_PROTO_TYPE macro which is intended
to be removed when the OSS deps are synchronized.
This will automatically cause a build break and force us
to revert this diff.

Reviewed By: @billf

Differential Revision: D2356336