Remove a set of defines for windows in chronoutils.h
Summary: I don't know where this originally came from, but, while it might have worked in whatever project it was used in at the time, defining things like this just breaks things. (it even has the `mode_t` define wrong, `mode_t` is supposed to be a typedef to `short`)
This removes the offending block of code.
Closes https://github.com/facebook/hhvm/pull/5952

Reviewed By: @sgolemon

Differential Revision: D2352248