Support first class enums in as/disas
Summary: Add support for first class enums in as/disas. Mostly this means adding a .enum_ty directive to specify the enumBaseTy field in PreClass. To support this we add a modified version of the existing type constraint syntax that only has the type constraint part and not the "user type" part.

Reviewed By: @paulbiss

Differential Revision: D2355027