Allow overriding the location to find the default php.ini
Summary: Right now, this is hardcoded to `/etc/hhvm/php.ini`, but that's
not always the right location, e.g., with brew on OS X. This allows that
default directory to be overridden.

This was attempted a while ago, with D1505433, but it did it the wrong
way and was reverted as D1560037. We don't want to use the install dir
as a prefix, leading to things like `/usr/etc`, but rather allow it to
be overridden entirely.

Reviewed By: @paulbiss

Differential Revision: D2358603