Exclude util/light-process.cpp under MSVC
Summary: LightProcess is heavily dependent upon fork, which is not available under Windows, so I've taken the route of simply not porting it to MSVC, and re-writing any required code that uses it, to not use it.
This simply removes the sources for them from the build under MSVC, and it shouldn't effect any other platforms.
Closes https://github.com/facebook/hhvm/pull/5944

Reviewed By: @sgolemon

Differential Revision: D2352172

Pulled By: @sgolemon