fix three problems in new strtr algorithm
When the input string is shorter than the shortest input
pattern the last position to visit is set to a negative, which
due to the unsigned nature rolls over to a very big number.
Fix short circuits and returns input string when all patterns
are longer than the input string.

The LRU cache is holding onto StringData between requests through
the PatAndRepl class, but StringData uses the request-local
allocator. To ensure the patterns and replacements survive
across requests they are now copied into a std::string.

Additionally this included a potential race condition under
concurrency. initTables() could be running in one thread when
the WuManberReplacement is pulled from the LRU cache.  This
second execution could start running the replacement before
initTables() has completed. The class has been adjusted so
it can be declared const at construction and does not change
its internal state.