Revert 'Implements `hh_server --daemon` on Windows.'
Summary: D2342443 reworked how hh_server daemonizes itself and in the process caused the hh_server monitor_daemon process to fail to close stdout/stderr. This caused trouble with processes that run `hh_client` and read the output, since its stdout is held open by the process running `monitor_daemon`.

Daemonizing a process is subtle and Daemon.fork does a good job of centralizing it all into one place and (probably) getting it right. I think that we ought to keep using it instead of opencoding it here.

Reviewed By: @int3

Differential Revision: D2363291