Remove maximum file storage limit
Summary: If we try to prime more than this limit we'll just prime straight to APC, regardless of whether the data is used. There's no reason to have this limit so this diff removes it. It retains the "Full" StorageState, reachable only if we can't add more files to the APC on-disk cache.

Will having a FileStorage.MaxSize config option in the config file affect this or will we just ignore anything unrecognised?

Reviewed By: @edwinsmith, @markw65

Differential Revision: D2360365