expand typedef for aliases when needed
Summary: This is mostly for perflab results.

I would like to carry the generic type variables into the runtime for type aliases, so that we can expand them when necessary. It's stored as a static string (comma separated names) so I don't expect too much of a regression.

===
This diff stores the generic typevar names for aliases as comma separated strings

  newtype MyAlias<Tk, Tv> = (function (Tk): Tv);
  class C {
    const type T = MyAlias<int, string>;
  }

The TypeStructure for MyAlias contains a 'typevars' field which is the string 'Tk,Tv'. When we resolve type constant C::T, we can obtain the type structure for MyAlias, which contains type variables Tk and Tv, and we substitute in the resolved actual generic types (TypeStructure of int, and that of string). This diff does not perform the substitution yet.

More details can be found here: https://quip.com/pQZ6AaRKNbD3 towards the end.

Reviewed By: @paulbiss

Differential Revision: D2361950