Implements file watcher on Windows.
This a simplifed version of Orvid patch (D39327).

We create a new thread for every roots directory to be watched. This
thread loops on the function `ReadDirectoryChangesW` and store
received events in a lock-less shared list. Whenever a new events is
inserted in the shared list, this thread also pushes a single
character in a given pipe. On the OCaml side, we might wait for new
event with `Unix.select` ober the same pipe.