Cleaner exit when a light process dies
Summary: When a light process dies, we attempt to stop the server, and exit
cleanly. Part of the exit process is to shut down the light processes;
when it attempts to shut down the one that was killed, it throws an
exception that isn't caught, resulting in a call to std::terminate.

At the time we shut down the lightprocesses, we haven't destroyed the
folly::Singletons, and there could still be server threads (and other
threads) running. At this point, shutdown is very likely to fail.

This diff wraps all the light process calls in a try/catch so that we
can just log a warning, and return a suitable error value.

There are still a few issues to resolve; we start the server shutdown
from a signal handler, and the code involves mutexes and condition
variables, so its technically undefined behavior.  For the most part,
this seems to have fixed the crashes though.

Reviewed By: @jano

Differential Revision: D2368228