Do not root dynPropTable
Summary: This table holds the property array for any object with dynamic
properties. When objects are freed, entries in the table are also
freed. It should not be a root for GC purposes; this code takes
it out of the root set and adds logic to update the table when
objects are swept.

Reviewed By: @ricklavoie

Differential Revision: D2365996