Vasmify f{call,uncBody}HelperThunk; rewrite former's API
Summary: In addition to the vasm rewrite, this changes the API around
fcallHelperThunk and fcallHelper somewhat.  This shouldn't change
behavior or performance, but is a bit easier to reason about (and a bit
less architecture-dependent).

Some outdated comments in the old emitter described the need to pop and
push the native saved RIP into rvmfp()[AROFF(m_savedRip)], because
doFCall() modifies it.  This no longer appears to be the case---setting
m_savedRip happens in iopFCall() in the interpreter (outside doFCall),
and inside the func prologue in the TC (which fcallHelperThunk
dispatches to).

Instead, we can just align the native stack normally (by shifting rsp),
and do a ret in the intercepted case.

Reviewed By: @markw65

Differential Revision: D2355316