Port ext_string to MSVC
Summary: `qsort_r` is `qsort_s` under MSVC, and the context pointer is the first argument to the callback rather than the last.
I didn't want to deal with porting the multi-dimentional C99 stack allocated array to use alloca, so I just disabled that for now.
We don't have `nl_langinfo` under MSVC, so raise a warning when it's called.
Closes https://github.com/facebook/hhvm/pull/6009

Reviewed By: @sgolemon

Differential Revision: D2367480