UVS: static member access
Summary: Adds new production rules for static member access now allowed
in PHP7, similar to D2370455. Also adds a test pulled from the UVS PR.

Two bits of the test are commented out. The second one relies on a
precedence change that breaks BC and we haven't done yet, diffs
forthcoming for that. The first is due to an inconsistency in the PHP7
grammar: you're allowed to do `var_dump(['array', 'literal'][0]::$x)`
and `new $y::$x`, but not `new ['array', 'literal'][0]::$x`. PHP7 throws
a parse error on the latter, but due to the construction of our parser,
trying to support the former leads to reduce/reduce conflicts due to the
latter. Could probably work around this, but didn't seem worth it.

Part of https://github.com/facebook/hhvm/issues/5523

Reviewed By: @paulbiss

Differential Revision: D2371495