Don't allow classes/aliases to match inbuilt types
It's a fatal error in the run time to have a class or type alias with
the same name as one of the primitive type hints, so don't allow them in
the type checker either.

Fixes #5865

Test Plan: Lots of new tests. Two for each inbuilt type that can't be
matched (one in a namespace, one not, and with different cases).