Don't use auto const when assigning a couple of bitsets
Summary: In today's issue of MSVC regressions and oddities, I present to you the im-a-const construction dilema!

Basically, for some absurd reason, MSVC seems to think the type of the result of `ALocBits{meta->conflicts}` is a `const`.

The RC version actually handled this just fine; this issue wasn't present until the first actual release of VS 2015.
And yes, these really are the only two places this change is needed.
Closes https://github.com/facebook/hhvm/pull/6108

Reviewed By: @sgolemon

Differential Revision: D2373253