Fix findPredTransIDs in LoopAnalysis
Summary: This function may reach the entry block in the region, which is
generated upfront and doesn't contain a proper profTransID.  This
triggers an assert in debug builds.  I'm not sure why we didn't hit
this in the original diff, but I started seeing it apparently after
the inlining work moved to translateRegion.

Reviewed By: @paulbiss

Differential Revision: D2372535