Clean up opPre and propPre in member-operations
Summary: opPre is the same thing as tvToCell, but without the LIKELY attribute.  propPre
handles stuff like null getting promoted to stdClass for defining prop
operations, so that was simplified.  For IssetEmptyProp it was mainly just being
used to check if the base was an object, so just do that inline instead.

Reviewed By: @swtaarrs

Differential Revision: D2365891