UVS: allow object dereference on more expressions
Summary: Conceptually not hard -- add a few productions to allow a deref
on more types of expressions in parenthesis. In practice, factoring this
nicely such that Bison would allow it leads to shift/reduce conflicts,
leading to this kind of ugly copy-paste.

Part of https://github.com/facebook/hhvm/issues/5523

Reviewed By: @paulbiss

Differential Revision: D2376013