Null out pointers when destroying arrays in debug mode
Summary: The current method of destroying arrays can cause assertions if the GC
runs during the destruction. The reason is that we've freed some of the elements
of the array, but not modified the pointers within the array. Since this array
is still reachable, the GC will attempt to trace it, and assert as soon as it
sees a reachable pointer to freed memory. We can avoid this by pro-actively
nulling out the pointers in the array as we free them.

Since this has a performance impact, and only really matters in debug mode
(since we wouldn't assert in a prod build), only do this nulling in debug mode.

Reviewed By: @edwinsmith

Differential Revision: D2374870