Support new minstrs in the jit
Summary: This diff adds complete but suboptimal jit support for the new
minstrs. The initial code that comes out of irgen looks a lot worse than what
we generate for the current minstrs, and we're depending on load/store-elim to
clean it up. One nice side-effect of this is that the new code in
irgen-minstr.cpp is more straightforward; we don't have to deal with anything
that used to walk the whole member vector, like checkMIState() or
nLogicalRatchets.

There are a couple major performance issues left to be fixed in later diffs:
- No support for specialized property access (this isn't very meaningful until
  we have hhbbc support).
- memory-effects.cpp has been made more conservative around MInstrState because
  it can contain values between bytecodes now. It should be correct but still
  needs some tuning to properly clean up all the junk we emit for the new
  minstrs.

Reviewed By: @alexmalyshev

Differential Revision: D2391994