Disable the takeover agent under MSVC
Summary: This is done by having all of it's functions return errors.
This is done this way because it is far far cleaner than excluding the takeover agent completely from the build, which would require all server implementations to have pieces #ifdef'd out, along with several other places as well, resulting in nothing but a mess.
Closes https://github.com/facebook/hhvm/pull/6125

Reviewed By: @sgolemon

Differential Revision: D2375978