Add support for interior pointers in conservative GC scanner
Summary: Add support for interior pointers to the GC conservative scanner. Any
pointer which points at any location within an allocation will mark that
allocation as reachable.

This is implemented using a sorted vector of regions (base pointer and
length). A binary search is done to find the first region with a base pointer
greater than the pointer being considered. The region preceeding this one is the
only possible one which can contain the pointer, so it is checked if the pointer
lies within that region's boundaries.

Reviewed By: @edwinsmith

Differential Revision: D2374822