Register active exceptions as GC roots
Summary: Exceptions can contain request-heap allocations, so must be scanned by
the GC. However, they are usually allocated on the system heap, so the GC will
not trace them, conservative or not. In addition, when an exception is thrown
and the stack is unwinding, it is stored somewhere within the runtime that we
don't have access to.

So, explicitly treat any active exception as a GC root. All the
ExtendedException constructors register themselves with the memory manager as
roots, and unregister themselves in their destructor. The GC will explicitly
trace the exceptions on this list.

Reviewed By: @edwinsmith

Differential Revision: D2379056