freeClone should smash prolog guards
Summary: When we destroy a Func, we smash its prologue guards so that another
Func allocated at the same address won't accidentally match.

We need to do the same thing for freeClone - because its even more
likely that we'll call the same guard (and pass it) in this case.

Noticed while doing this that we also don't free up the tc used by the
Func when tc-recyling is enabled.

Reviewed By: @mxw

Differential Revision: D2387124