Initial support for new member instructions
Summary: This diff introduces the first set of new member bytecodes, along with
supporting infrastructure. I've added enough to support any CGetM with a local
or $this base. The new instructions are heavily suboped, partially to save
bytecode space but also because a lot of these operations have the same
signatures and very similar implementations. It might make sense to subop them
even with infinite bytecode space.

The main missing parts are that the JIT only knows enough about them to
correctly interpOne them and HHBBC support is nonexistent, both of which can
come in later diffs.

Each member instruction has been broken up into three main parts:
- A Base* instruction (BaseL, BaseH, ...). This corresponds to the location
  code from the member vector.
- Any number of intermediate Dim* instructions, which cover elements of the
  member vector before the final operation.
- A final operation. This diff adds QueryM* which handles CGetM, IssetM, and
  EmptyM; a few more will be added for SetM and the other minstrs.

If this three-way split sounds familiar, it's because it already exists in the
implementation of the current member instructions, both in the interpreter and
the jit.

One rough edge is that we need to keep any stack inputs on the stack until the
end of the final operation, so all final operations take an IVA immediate
indicating how many stack elements they should pop. These stack elements are
all a new flavor, CR, which could be a Cell or a return value. The current
member instructions can get the exact flavor of each element by reading through
the member vector, but not having the exact flavor should have no practical
downsides. We could replace all stack inputs to the member operations with
unnamed locals and clean this up a little but it's not necessary yet.

I've also added some tracing and did some minor cleanup in the emitter along
the way, to make debugging easier. I could probably be convinced to make those
a separate diff. And since I added the first bytecodes with more than one
subop, all the subop members in hhbbc instructions now have index suffixes.

Reviewed By: @alexmalyshev

Differential Revision: D2328347