Register RequestEventHandlers immediately
Summary: RequestEventHandlers are created, have their requestInit() callbacks
called, and then if an exception is not thrown, are registered with the
execution-context. This poses a problem for the GC as a collection can be
triggered during the requestInit() call, and the RequestEventHandler will not be
traced. Even conservative tracing of the stack will not help us, as
RequestEventHandlers cannot be allocated from the request local heap, and we
need explicit scanning of the execution-context's list to find them.

Fix this by registering them with the execution-context before calling
requestInit(). This leaves the issue of what to do if requestInit() throws. In
that case, we remove it from the execution-context's list by nulling out the
entry. When we go to iterate over the list, we ignore null entries. This is more
efficient than removing it from the list directly and possibly shifting over
lots of entries.

Reviewed By: @edwinsmith

Differential Revision: D2379958