Bump Third Party SHA and remove proxygen secure transport workaround
Summary: I updated hhvm-third-party to get the latest `thrift`, `proxygen`,
`wangle`, `squangle`, `mcrouter` and `folly` tags.

D2268791 was an interesting diff that required a couple extra
pushes to third-party.

Thrift was updated so now we can get rid of the workaround in
D2356336 that was causing the OSS build to fail without it.

Reviewed By: @jwatzman

Differential Revision: D2387115