Add support for building with precompiled headers under MSVC
Summary: This results in a significant decrease in the amount of time it takes to build hphp_runtime_ext and hphp_runtime_static. For hphp_runtime_static, it reduced build time for my from ~30 minutes to ~12 minutes. More could be done to improve the time even more, but these are good enough for now.
While technically dependent upon [folly#287](https://github.com/facebook/folly/pull/287), HHVM doesn't currently compile under MSVC to begin with, and this won't break the normal builds even if this is merged before that is merged upstream.
Closes https://github.com/facebook/hhvm/pull/6141

Reviewed By: @sgolemon

Differential Revision: D2379138

Conflicts:
	hphp/public_tld/CMake/HPHPCompiler.cmake

Pulled By: @sgolemon