Add more kind assertions to catch GC corruption more quickly
Summary: If the GC collects an allocation which is still reachable, it can
manifest itself in many possible ways. One way is that since a free'd allocation
has its size stored in the same place as where the count is normally stored, a
dec-ref can corrupt the heap.

To catch these kinds of issues faster, add more assertion checking that an
allocation's kind is correct. Namely, assert the kind is valid before doing any
ref-count manipulation, or before dispatching to various array
functions. Finally, when in debug mode, use a slightly slower version of
TV_GENERIC_DISPATCH which preserves all the assertions for the correct type.

Reviewed By: @edwinsmith

Differential Revision: D2384064