Don't defer updating position-limit inside php_map()
Summary: BaseMap::php_map() has an optimization where it only updates the
contained array's position-limit to the final value when the loop
exits. However, this is not compatible with GC, as this value tells the GC how
many elements to scan within the contained array. By only updating it at the end
of the loop, if we trigger a GC within the loop, the values we've processed so
far will not be traced. Fix this by incrementing the position-limit on every
trip through the loop.

Reviewed By: @edwinsmith

Differential Revision: D2380712