Remove tvScratch
Summary: It was meant as an optimization.  We would store member-inst bases in tvScratch
when we knew they weren't refcounted values.  However it doesn't end up being a
win because we don't distinguish IR instructions that just use tvScratch, versus
those that use all of MInstrState.  Thus instruction that would use tvScratch
would still be initializing and decreffing tvRef{1,2}.

This diff removes it, and replaces its uses with tvRef.

Reviewed By: @swtaarrs

Differential Revision: D2376930