Fix race setting up server sockets
Summary: Apparently, you can't bind to a port that another process is listening
on, but two processes can both bind to a port that neither is
listening on.

As a result, we have a race. Two instances can both bind the server
socket, but then they try to listen in the event loop; one will fail
and throw an exception, causing runInEventBaseThread to abort.

The docs for AsyncServerSocket say that listen must be called from the
event base thread, which would make fixing this awkward. But they same
the same about bind, and we've been running that directly from
ProxygenServer::start without issues. Code inspection (and common
sense) shows that the condition is actually that it must be run from
the event base thread, or the event base thread must not have started
yet. So just pull the listen calls out of the runInEventBaseThread.

Reviewed By: @afrind

Differential Revision: D2390877