Don't pass MInstrState* to jit helpers
Summary: The helpers that were taking an MInstrState* really only want a
pointer to tvRef. This simplifies things by removing the DefMIStateBase
instruction and just passing tvRef to the helpers that need it. It also greatly
improves what load/store-elim can do with the ratchetRefs code for new minstrs,
since we no longer think every helper can read/write all of MInstrState (it may
help current minstrs a little, if checkMIState() is missing some cases).

I also extended AMIState to hold a bitvector representing the different parts
of MInstrState, rather than just a single offset. I ended up not using it as
originally planned but it won't hurt anything and may be used in the future so
I left it in.

Reviewed By: @markw65

Differential Revision: D2395762