Kill TranslateRetryContext.inlines
Summary: This caching seems a premature optimization, and it doesn't work in
general.  I don't think this is a problem in trunk right now, but I
hit this with a diff I'm working on, and I think it can also cause
performance issues with profile-based regions.  The basic problem is
that the callee region can vary depending on the calling context (base
object and types of args).  In the particular case triggered by my
diff, the base object of a method call can vary, and so can the callee
function.  So we end up using a cached callee region for the wrong
callee function.  With PGO-based regions, I think we can end up using
a cached callee region for arg types that don't match (thus triggering
side exits).

So, for now, I'm removing this cache.  If necessary, we can
reintroduce it later and make it right by also using the calling
context in the key.

Reviewed By: @paulbiss

Differential Revision: D2393375