Remove tvScratch, take 2
Summary: It was meant as an optimization.  We would store member-inst bases in tvScratch
when we knew they weren't refcounted values.  However it doesn't end up being a
win because we don't distinguish IR instructions that just use tvScratch, versus
those that use all of MInstrState.  Thus instruction that would use tvScratch
would still be initializing and decreffing tvRef{1,2}.

This diff removes it, and replaces its uses with tvRef.

This first went in as D2376930 but was reverted. The issue was that we were
letting a pointer to the tempBase local in memberHelperPre() escape its
frame. tempBase has been moved to MemberState with the other similar
TypedValues; this diff is otherwise identical to D2376930.

Reviewed By: @markw65

Differential Revision: D2393320