Fix build with freetype in nonstandard location
Summary: Such as OS X. This would cause CMake to look for
`FindLibFreetype.cmake` which doesn't exist, we have
`FindFreetype.cmake`. Generates tremendous spew (unclear how it was
missed when testing the blamerev...) but happens to work if freetype is
in one of the directories already on your include path.

Reviewed By: @paulbiss

Differential Revision: D2398272