Updating hphp test runner for usability
Summary: In learning hhvm development, I found that the test script had some surprising
behavior. This came up when I mis-specified a test target, causing *all* tests
to be run, rather than getting an error that the test target was wrong. These
changes make the test script more intuitive/usable.

Reject unknown options rather than treating them as test targets. To use a test
target with a name starting with '-' either prefix with './' or use the '-'
option to treat all later arguments as test targets.

Reject unknown and not-found test targets, whether offered by the user or caused
by a broken convenience name.

Stop double-globbing test targets. The shell does globbing; globbing within the
application is necessary only for convenience targets and should not be applied
to targets that are *not* convenience targets.

Reviewed By: @ricklavoie

Differential Revision: D2392671