Filter dfind updates that don't belong under the .hhconfig dir
Summary: Relative_path.create raises an assertion if the given path does not
have the .hhconfig root dir as a prefix. However, this case is indeed possible
when dealing with updates from dfind, because it follows symlinks. So we filter
out those files that are not part of the .hhconfig-containing directory.

(The fact that it does that when the find service doesn't is kind of clowny,
and it would be nice to make their behavior uniform too, but I don't know the
dfind code well and having the check in ServerMain makes things more robust
anyway.)

Reviewed By: @​achow

Differential Revision: D2395047