Check for a nullptr XDebugServer in XDebugHookHandler::onOpcode
Summary: We can detach the server early in the request, such as if xdebug hits a
breakpoint and loses the connection to the client.  We leave the
XDebugHookHandler attached though (not really sure why...), and when it runs the
onOpcode hook it crashes trying to access the nullptr XDebugServer.

Other methods in XDebugHookHandler already check for a nullptr XDebugServer, so
I figured I'd do that instead of trying to detach the XDebugHookHandler as well.

Reviewed By: @paulbiss

Differential Revision: D2404028