Fix `input_value: bad object` in build
Summary: This is probably one of my better one-line fixes...

`input_value: bad object` is the result of the Marshal module trying to
unmarshal something that doesn't look like a marshalled OCaml object. Some
investigation revealed that it was failing to unmarshal the string "ready",
which is sent during the server startup phase, on an entirely different channel
altogether.

How does data from one channel find itself being written to another, and why
does using close_out instead of Unix.close fix things? Well, OCaml channels are
buffered, and Linux reuses file descriptors. So `Unix.close` would close a file
descriptor while retaining the OCaml channel buffer, and then Socket.accept
would reuse that fd value for the client connection. At some point we would
flush the old buffer, causing the client to read the stale data.

One might notice that wakeup_client calls `flush` immediately after writing its
value, so the OCaml buffer should get emptied... unless the client isn't trying
to read that value (which is true when doing build). So an exception gets
thrown, which we ignore, and the OCaml runtime keeps around the buffer that it
failed to flush (which is probably a shortcoming on its part).

This regression was introduced in {D2367655}.

For the curious, the OCaml channel code can be found at
https://github.com/ocaml/ocaml/blob/ac62e93d6b61ae76e18dee530b7f3969ef690219/byterun/io.c

Reviewed By: @jwatzman

Differential Revision: D2405802