Support new minstrs in hhbbc
Summary: There's no new analysis here; most of the changes are some light
refactoring so the existing logic can be shared by both new and old
minstrs. Two of the members of MIS have been moved to State since they can live
across bytecodes now. The remaining members aren't relevant to the new
minstrs.

I also rearranged the in() functions in interp.cpp so they don't all have to be
defined in the same .cpp, which cleans things up a little (and will help even
more when we add another 20 new minstrs). And the verifier needed some tweaking
to allow AssertRAT* bytecodes in the middle of a member operation.

Reviewed By: @alexmalyshev

Differential Revision: D2409548