Be compatible with OCaml-4.03 (unreleased) (fix #6120)
Summary: Futur version of OCaml will rename 'uint64' (as defined in `<caml/config.h>` usually imported through `<caml/mlvalues.h>`) into `uint64_t` in order to stick with typedef introduced in C99.

Meanwhile, the MinGW compiler when called by `ocamlopt` is not called with `-std=c99` and the type `uint64_t` is not defined. This patch partially revert #5680 and introduce a typedef specific to MinGW.
Closes https://github.com/facebook/hhvm/pull/6149

Reviewed By: @JoelMarcey

Differential Revision: D2388554