Centralize encoding/decoding of bytecode Ops
Summary: All reading and writing of Ops from/to a bytecode stream now happens
in two functions in hhbc-codec.h. This allows us to easily change the encoding
to support more than 256 Ops, which I've done (details in hhbc-codec.h).

Most of the changes are to replace direct dereferencing of an Op* with calls to
decode_op() or peek_op(). I also replaced all uses of `const Op*` with PC
(which is `const uint8_t*`) - now that the bytecode stream doesn't actually
contain valid Op values it doesn't make sense to pass around pointers
pretending it does (and it was always a little sketchy, since the Op* we passed
around often pointed to immediates rather than Ops).

Reviewed By: @edwinsmith

Differential Revision: D2402607