Cleanup m_pendingException safely
Summary: clearPendingException destroys a backtrace array, which might contain
objects, which might have destructors, which would cause code to run,
which would notice that there's a pending exception, which would try
to throw it, even though its partially destructed.

This has always been a bug, and we would double free the exception,
but now that we keep an index into the m_exceptionRoots vector in the
ExtendedException, we would crash trying to remove it from the vector
a second time (after its index had been cleared).

Reviewed By: @swtaarrs

Differential Revision: D2422895