Add an override for a hack
Summary: We have a hack from years ago which always decodes post data as if it
were application/x-www-form-urlencoded, regardless of the Content-type
header (some clients weren't sending the header, but expected hhvm to
decode them anyway). This can cause issues for some endpoints - eg that
take binary data - since they might happen to decode parameters that the
endpoint understands.

Allow opting out on a per virtual host, or per virtual host/endpoint
basis.

Long term, we'll try to determine if the hack is still needed, and remove
it if not.

Reviewed By: @alexmalyshev

Differential Revision: D2426568